% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{color_palette}
\alias{color_palette}
\title{getting distinguishable colours for clusters}
\usage{
color_palette(num_col = 40)
}
\description{
getting distinguishable colours for clusters
}
\details{
ref: http://stackoverflow.com/questions/15282580/how-to-generate-a-number-of-most-distinctive-colors-in-r
}
