% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{compute_lfsr}
\alias{compute_lfsr}
\title{compute local false sign rate (lfsr)}
\usage{
compute_lfsr(samples)
}
\arguments{
\item{samples}{numeric vector of posterior samples}
}
\value{
numeric between 0 and 0.5
}
\description{
compute local false sign rate (lfsr)
}
