% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_lfsr_from_csv}
\alias{compute_lfsr_from_csv}
\title{Compute local false discovery rate (lfdr) and local false sign rate (lfsr) for mu/beta1, adjusted by estimated null distribution}
\usage{
compute_lfsr_from_csv(
  csv_files,
  par_pattern,
  stan_variables,
  mu_ntc_col = NULL,
  tau_ntc_col = NULL
)
}
\arguments{
\item{samples}{numeric vector of posterior samples}
}
\value{
numeric between 0 and 1
}
\description{
Compute local false discovery rate (lfdr) and local false sign rate (lfsr) for mu/beta1, adjusted by estimated null distribution
}
