% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_rope_lfdr}
\alias{compute_rope_lfdr}
\title{compute local false discovery rate (lfdr) for mu/beta1, adjusted by estimated null distribution}
\usage{
compute_rope_lfdr(mu_g, mu_ntc, tau)
}
\arguments{
\item{samples}{numeric vector of posterior samples}
}
\value{
numeric between 0 and 1
}
\description{
compute local false discovery rate (lfdr) for mu/beta1, adjusted by estimated null distribution
}
