% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_size_factors}
\alias{compute_size_factors}
\title{Calculate size factors via median-of-ratios (ref. DESeq2)}
\usage{
compute_size_factors(dough)
}
\arguments{
\item{dough}{dough object with $data$counts}
}
\value{
size factors
}
\description{
Calculate size factors via median-of-ratios (ref. DESeq2)
}
