% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confusion_matrix_stitch_parallel_MAP.R
\name{Confusion_matrix_stitch_parallel_MAP}
\alias{Confusion_matrix_stitch_parallel_MAP}
\title{Construct the global Confusion matrix.}
\usage{
Confusion_matrix_stitch_parallel_MAP(pip)
}
\description{
Construct the global Confusion matrix.
}
