% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{extract_parameters}
\alias{extract_parameters}
\title{extract posterior draws for one parameter}
\usage{
extract_parameters(biscuit, pars)
}
\arguments{
\item{biscuit}{fitted biscuit object}

\item{pars}{single parameter base name (e.g. "mu")}
}
\value{
dataframe with columns: draw, parameter index, value
}
\description{
extract posterior draws for one parameter
}
