% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscuit.R
\name{fit_biscuit}
\alias{fit_biscuit}
\title{Fit the biscuit Stan model}
\usage{
fit_biscuit(
  dough,
  output_dir,
  filter = TRUE,
  save_samples = TRUE,
  n_parallel_chains = 4,
  seed = 13,
  pseudocount = TRUE
)
}
\arguments{
\item{dough}{dough object with data stored in \code{$data}}

\item{output_dir}{directory to save biscuit output files}

\item{filter}{logical; whether to filter counts before fitting}

\item{save_samples}{logical; whether to save posterior draws}

\item{n_parallel_chains}{integer; number of chains to run in parallel}

\item{seed}{integer; seed for reproducibility}

\item{pseudocount}{logical; whether to add a pseudocount to the count matrix}
}
\value{
biscuit object with \code{$data} and \code{$fit}
}
\description{
Fit the biscuit Stan model
}
