% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscuit.R
\name{generate_biscuit_input}
\alias{generate_biscuit_input}
\title{Generate model input data for the biscuit Stan model}
\usage{
generate_biscuit_input(dough, pseudocount = TRUE)
}
\arguments{
\item{dough}{dough object with \code{$data}}

\item{pseudocount}{logical; whether to add a pseudocount of 1 to the count matrix}
}
\value{
named list of input data for the biscuit Stan model
}
\description{
Generate model input data for the biscuit Stan model
}
