% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_norm_counts}
\alias{get_norm_counts}
\title{normalize counts}
\usage{
get_norm_counts(dough)
}
\arguments{
\item{dough}{dough (or biscuit) object with $data$counts}
}
\value{
norm_counts matrix of normalized read counts, via median of ratios method, using DESeq2
}
\description{
normalize counts
}
