% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMM_Gibbs_MCMC_parallel.R
\name{IMM_Gibbs_MCMC_parallel}
\alias{IMM_Gibbs_MCMC_parallel}
\title{Main MCMC Engine #######################}
\usage{
IMM_Gibbs_MCMC_parallel(pip)
}
\description{
Main MCMC engine. Do not change anything. This runs in parallel 
   where each parallel run takes in a matrix X of all cells and a 
   gene batch i.e. dim(X) is numcells x gene_batch.
}
\details{
DO NOT change anything in IMM.MCMC()
}
