% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfbiscuit.R, R/biscuit.R
\name{knead_dough}
\alias{knead_dough}
\title{generate model input data, intended for internal use (wrapper for generate_biscuit_input)}
\usage{
knead_dough(dough, pseudocount = TRUE)

knead_dough(dough, pseudocount = TRUE)
}
\arguments{
\item{dough}{dough object, with data stored in $data}

\item{pseudocount}{logical indicating whether a pseudocount should be added to the count matrix}
}
\value{
list of input data used to fit the biscuit model

list of input data used to fit the biscuit model
}
\description{
generate model input data, intended for internal use (wrapper for generate_biscuit_input)

generate model input data, intended for internal use (wrapper for generate_biscuit_input)
}
