% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{loadSingleCellMatrix}
\alias{loadSingleCellMatrix}
\title{Load matrix for single cell data}
\usage{
loadSingleCellMatrix()
}
\description{
Load matrix for single cell data
}
\details{
this assumes the input data has both column and row names.
}
