% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_count_logfc}
\alias{plot_count_logfc}
\title{plot MA of guides using empirical estimates (treatment vs control)}
\usage{
plot_count_logfc(dough)
}
\arguments{
\item{dough}{dough object with $counts}
}
\value{
p ggplot object
}
\description{
plot MA of guides using empirical estimates (treatment vs control)
}
