% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_counts_distribution}
\alias{plot_counts_distribution}
\title{plot histogram of read counts per sample}
\usage{
plot_counts_distribution(dough, normalized = TRUE)
}
\arguments{
\item{dough}{object with $data$counts}

\item{normalized}{logical indicating whether to normalize counts or not}
}
\value{
p ggplot object
}
\description{
plot histogram of read counts per sample
}
