% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_counts_violin}
\alias{plot_counts_violin}
\title{plot distribution of log2 (optional: normalized) counts per sample}
\usage{
plot_counts_violin(dough, normalize = TRUE)
}
\arguments{
\item{dough}{dough object with $data$counts}

\item{normalized}{logical indicating whether to normalize counts or not}
}
\value{
p ggplot object
}
\description{
plot distribution of log2 (optional: normalized) counts per sample
}
