% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dispersion}
\alias{plot_dispersion}
\title{plot moment estimated dispersion}
\usage{
plot_dispersion(dough)
}
\arguments{
\item{dough}{dough object with $data$counts}
}
\value{
p ggplot object
}
\description{
plot moment estimated dispersion
}
