% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_eff_violin_gene}
\alias{plot_eff_violin_gene}
\title{plot stacked density plot of phi, and corresponding phi times gamma, visualizing variation in dispersion across samples}
\usage{
plot_eff_violin_gene(biscuit, gene_name)

plot_eff_violin_gene(biscuit, gene_name)
}
\arguments{
\item{biscuit}{a biscuit object with $fit, $results}

\item{gene_name}{name of gene}
}
\value{
ggplot object

p ggplot object
}
\description{
plot stacked density plot of phi, and corresponding phi times gamma, visualizing variation in dispersion across samples

plot violin plot of posterior distribution of efficiency for guide that target a given gene
}
