% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gene_rank}
\alias{plot_gene_rank}
\title{plot score versus rank plot of genes}
\usage{
plot_gene_rank(biscuit, lfdr_threshold = 0.05, top_n = 10)
}
\arguments{
\item{biscuit}{a biscuit object with $results}

\item{lfdr_threshold}{threshold to highlight significant genes}

\item{top_n}{number of top genes to label}
}
\value{
ggplot object
}
\description{
plot score versus rank plot of genes
}
