% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gene_volcano}
\alias{plot_gene_volcano}
\title{plot volcano plot for inferred gene-level effects}
\usage{
plot_gene_volcano(biscuit, lfdr_threshold = 0.05, top_n = 10)
}
\arguments{
\item{biscuit}{a biscuit object with $results}

\item{lfdr_threshold}{threshold to highlight significant genes}

\item{top_n}{number of top genes to label}
}
\value{
ggplot object
}
\description{
plot volcano plot for inferred gene-level effects
}
