% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_guide_density}
\alias{plot_guide_density}
\title{plot densities of targeting vs non-targeting guide log fold changes}
\usage{
plot_guide_density(biscuit)
}
\arguments{
\item{biscuit}{a fitted biscuit object with $results, $data}
}
\value{
p ggplot object
}
\description{
plot densities of targeting vs non-targeting guide log fold changes
}
