% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_guide_violin}
\alias{plot_guide_violin}
\title{plot violin plot of biscuit targeting vs non-targeting control guide log fold changes}
\usage{
plot_guide_violin(biscuit)
}
\arguments{
\item{biscuit}{fitted biscuit object with $results$beta1 and $data$controls}
}
\value{
ggplot object
}
\description{
plot violin plot of biscuit targeting vs non-targeting control guide log fold changes
}
