% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_guides_per_gene}
\alias{plot_guides_per_gene}
\title{plot histogram of number of guides per gene}
\usage{
plot_guides_per_gene(dough)
}
\arguments{
\item{dough}{dough object with $data$row_data}
}
\value{
ggplot object
}
\description{
plot histogram of number of guides per gene
}
