% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_logfc_density}
\alias{plot_logfc_density}
\title{plot density of empirical logFC of guides, startified by non-targeting and targeting guides}
\usage{
plot_logfc_density(dough)
}
\arguments{
\item{dough}{dough object with $data$counts}
}
\value{
p ggplot object
}
\description{
plot density of empirical logFC of guides, startified by non-targeting and targeting guides
}
