% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mu_beta1_density}
\alias{plot_mu_beta1_density}
\title{plot stacked plot of posterior densities of a given gene and the guides that target it}
\usage{
plot_mu_beta1_density(biscuit, gene_name)
}
\arguments{
\item{biscuit}{a biscuit object with $data, $fit}

\item{gene_name}{name of gene}
}
\value{
p ggplot object
}
\description{
plot stacked plot of posterior densities of a given gene and the guides that target it
}
