% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_phi_gamma_density}
\alias{plot_phi_gamma_density}
\title{plot stacked density plot of phi, and corresponding phi times gamma, visualizing variation in dispersion across samples}
\usage{
plot_phi_gamma_density(biscuit)
}
\arguments{
\item{biscuit}{a biscuit object with $results and $fit}
}
\value{
ggplot object
}
\description{
plot stacked density plot of phi, and corresponding phi times gamma, visualizing variation in dispersion across samples
}
