% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_sample_correlation}
\alias{plot_sample_correlation}
\title{plot sample correlation}
\usage{
plot_sample_correlation(dough)
}
\arguments{
\item{dough}{dough object with $counts}
}
\value{
p ggplot object
}
\description{
plot sample correlation
}
