% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_MCMC_genesplit_merge.R
\name{post_MCMC_genesplit_merge}
\alias{post_MCMC_genesplit_merge}
\title{Combine the runs/gene_batches}
\usage{
post_MCMC_genesplit_merge(pip)
}
\description{
Postprocess MCMC chains from multiple parallel runs
}
