% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISCUIT_main.R
\name{post_process}
\alias{post_process}
\title{Compute imputed data based on inferred variables}
\usage{
post_process(pip)
}
\description{
Compute imputed data based on inferred variables 
  and generate plots.
}
