% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BISCUIT_main.R
\name{singlecell_processor}
\alias{singlecell_processor}
\title{call BISCUIT}
\usage{
singlecell_processor(
  input_file_name = stop("missing the raw data table file!"),
  input_data_tab_delimited = TRUE,
  is_format_genes_cells = TRUE,
  choose_cells = 3000,
  choose_genes = 150,
  gene_batch = 50,
  num_iter = 20,
  num_cores = detectCores() - 4,
  z_true_labels_avl = TRUE,
  num_cells_batch = 1000,
  alpha = 1,
  output_folder_name = "./output"
)
}
\arguments{
\item{input_data_tab_delimited}{set to TRUE if the input data is tab-delimited}

\item{is_format_genes_cells}{set to TRUE if input data has rows as genes and columns as cells}

\item{choose_cells}{comment if you want all the cells to be considered}

\item{choose_genes}{comment if you want all the genes to be considered}

\item{gene_batch}{number of genes per batch, therefore num_batches = choose_genes (or numgenes)/gene_batch. Max value is 150}

\item{num_iter}{number of iterations, choose based on data size.}

\item{num_cores}{number of cores for parallel processing. Ensure that detectCores() > 1 for parallel processing to work, else set num_cores to 1.}

\item{z_true_labels_avl}{set this to TRUE if the true labels of cells are available, else set it to FALSE. If TRUE, ensure to populate 'z_true' with the true labels in 'BISCUIT_process_data.R'}

\item{num_cells_batch}{set this to 1000 if input number of cells is in the 1000s, else set it to 100.}

\item{alpha}{DPMM dispersion parameter. A higher value spins more clusters whereas a lower value spins lesser clusters.}

\item{output_folder_name}{give a name for your output folder.}
}
\description{
run single cell data pre-processing via \code{BISCUIT} algorithm.
}
