% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{summarize_parameters}
\alias{summarize_parameters}
\title{Summarize posterior parameters}
\usage{
summarize_parameters(
  dough,
  fit_summary,
  pars = c("mu", "beta1", "beta0", "phi", "eff", "mu_ntc", "tau_ntc")
)
}
\arguments{
\item{dough}{dough object with \code{$fit}}

\item{pars}{character vector of parameter names to summarize}
}
\value{
dough object with \code{$results} appended
}
\description{
Summarize posterior parameters
}
