% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{summarize_params}
\alias{summarize_params}
\title{summarize posterior draws}
\usage{
summarize_params(biscuit, pars = c("beta1", "mu", "beta0", "phi", "gamma"))
}
\arguments{
\item{biscuit}{fitted biscuit object}

\item{pars}{list parameters to summarize (e.g. mu, beta1)}
}
\value{
biscuit object with $results
}
\description{
summarize posterior draws
}
