% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscuit.R
\name{fit_biscuit}
\alias{fit_biscuit}
\title{runs biscuit}
\usage{
fit_biscuit(
  dough,
  output_dir,
  save_samples = TRUE,
  n_parallel_chains = 4,
  seed = 13,
  pseudocount = TRUE
)
}
\arguments{
\item{dough}{dough object, with data stored in $data}

\item{output_dir}{directory to save biscuit output files}

\item{save_samples}{logical to indicate}

\item{n_parallel_chains}{integer indicating the number chains to run in parallel}

\item{seed}{integer indicating the seed for reproducibility}

\item{pseudocount}{logical indicating whether a pseudocount should be added to the count matrix}
}
\value{
biscuit object, with $data and $fit
}
\description{
runs biscuit
}
