% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfbiscuit.R, R/biscuit.R
\name{generate_biscuit_input}
\alias{generate_biscuit_input}
\title{generate model input data, intended for internal use}
\usage{
generate_biscuit_input(dough, pseudocount = TRUE)

generate_biscuit_input(dough, pseudocount = TRUE)
}
\arguments{
\item{dough}{dough object with $data}

\item{pseudocount}{logical indicating whether a pseudocount should be added to the count matrix}
}
\value{
list of input data used to fit the biscuit Stan model

list of input data used to fit the biscuit Stan model
}
\description{
generate model input data, intended for internal use

generate model input data, intended for internal use
}
