% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{make_playdough}
\alias{make_playdough}
\title{simulate count data using either fixed, parametric, or empirical gene effects}
\usage{
make_playdough(
  n_genes,
  guides_per_gene,
  n_control = 2,
  n_treatment = 2,
  p_effects = 0.1,
  p_positive = 0.1,
  fold_change = 4,
  expression = c(4, 6),
  guide_efficiency = 0.8,
  guide_sd = 0.5,
  n_ntc = 1000,
  seed = 13,
  gene_params = list(mu = log(fold_change), sd = 2),
  effect_mode = c("fixed", "parametric", "empirical"),
  counts = NULL,
  quantiles = c(0.05, 0.95)
)
}
\arguments{
\item{n_genes}{number of genes to simulate}

\item{guides_per_gene}{number of guides per gene to simulate}

\item{n_control}{number of control samples}

\item{n_treatment}{number of treatment samples}

\item{p_effects}{proportion of genes with a non-zero effect}

\item{p_positive}{proportion of effects that are positive}

\item{fold_change}{absolute fold change of non-zero effects for fixed-effect mode, default mean for parametric mode}

\item{guide_sd}{optional standard deviation for guide effect simulation}

\item{n_ntc}{number of non-targeting control guides}

\item{seed}{integer seed for reproducibility}

\item{gene_params}{optional list for parametric mode (e.g. list(mu = log(4), sd = 0.2))}

\item{effect_mode}{"fixed", "parametric", or "empirical": modes that determine how gene effects are simulated}

\item{counts}{optional matrix of raw experimental counts; if provided, will use to calculate baseline expression and guide-wise dispersion}

\item{quantiles}{optional quantiles (e.g. c(0.05, 0.95)) to draw significant effects from in empirical mode}
}
\value{
list with simulated counts, guide to gene mapping, sample design and true parameter values
}
\description{
simulate count data using either fixed, parametric, or empirical gene effects
}
