% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{make_playdough}
\alias{make_playdough}
\title{simulate count data using moments estimated from real data}
\usage{
make_playdough(
  counts,
  n_genes,
  guides_per_gene,
  n_control = 2,
  n_treatment = 2,
  p_effects = 0.1,
  p_positive = 0.1,
  fold_change = 4,
  n_ntc = 1000,
  seed = 13
)
}
\arguments{
\item{counts}{matrix of raw experimental counts}

\item{n_genes}{number of genes to simulate}

\item{guides_per_gene}{number of guides per gene to simulate}

\item{n_control}{number of control samples}

\item{n_treatment}{number of treatment samples}

\item{p_effects}{proportion of effects}

\item{p_positive}{proportion of effects that are positive}

\item{fold_change}{absolute fold change of significant effects}

\item{n_ntc}{number of non-targeting control guides}

\item{seed}{numeric seed for reproducibility}
}
\value{
list including simulated counts, guide to gene mapping, sample design and true parameter values
}
\description{
simulate count data using moments estimated from real data
}
