% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_counts}
\alias{normalize_counts}
\title{normalize counts using the median-of-ratios method (ref. DESeq2)}
\usage{
normalize_counts(dough)
}
\arguments{
\item{dough}{dough (or biscuit) object with $data$counts}
}
\value{
matrix of normalized read counts
}
\description{
normalize counts using the median-of-ratios method (ref. DESeq2)
}
