% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_guide_ma}
\alias{plot_guide_ma}
\title{plot MA plot of guides using inferred guide effects (beta1)}
\usage{
plot_guide_ma(biscuit, lfsr_threshold = 0.05, top_n = 10)
}
\arguments{
\item{biscuit}{biscuit object with $results}

\item{lfsr_threshold}{threshold to highlight significant guides}

\item{top_n}{number of top guides to label}
}
\value{
ggplot object
}
\description{
plot MA plot of guides using inferred guide effects (beta1)
}
