% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mu_beta1}
\alias{plot_mu_beta1}
\title{plot density plot of posterior distribution of a given gene and posterior means of guides that target it}
\usage{
plot_mu_beta1(biscuit, gene_name, lfdr_threshold = 0.05)
}
\arguments{
\item{biscuit}{a biscuit object with $fit, $results}

\item{gene_name}{name of gene}

\item{lfsr_threshold}{threshold to highlight significant gene}
}
\value{
p ggplot object
}
\description{
plot density plot of posterior distribution of a given gene and posterior means of guides that target it
}
