% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{summarize_parameters}
\alias{summarize_parameters}
\title{summarize posterior draws}
\usage{
summarize_parameters(biscuit, pars = c("mu", "beta1", "beta0", "phi", "gamma"))
}
\arguments{
\item{biscuit}{biscuit object with $fit}

\item{pars}{list parameter names to summarize (e.g. c('mu', 'beta1'))}
}
\value{
biscuit object with $results
}
\description{
summarize posterior draws
}
