% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{trim_dough}
\alias{trim_dough}
\title{filter targeting guide counts}
\usage{
trim_dough(dough, min_counts = 30, min_guides_per_gene = 2, verbose = TRUE)
}
\arguments{
\item{dough}{dough object with $data: $data$counts, $data$row_data, $data$col_data (optional: $data$ntc)}

\item{min_counts}{minimum total counts per guide across samples to keep}

\item{min_guides_per_gene}{minimum number of guides per gene to keep the gene}

\item{verbose}{logical to print out dimensions before and after filtering}
}
\value{
dough object with filtered data
}
\description{
filter targeting guide counts
}
